+++ R/data.R
@@ -1,0 +1,13 @@
+#' DisMAP Regions
+#'
+#' A multiline simple feature boundary of regions and seasons used in DisMAP.
+#'
+#' @format ## `dm_regions`
+#' A data frame with 14 features and 4 fields:
+#' - `dataset_code`: DisMAP dataset code
+#' - `region`: Region name
+#' - `season`: Season name
+#' - `distribution_project_code`: DisMAP distribution project code
+#'
+#' @source <https://services2.arcgis.com/C8EMgrsFcRFL6LrL/ArcGIS/rest/services/DisMAP_Regions_CURRENT/FeatureServer/1>
+"dm_regions"