+++ R/dismapr.r
@@ -3,37 +3,35 @@
 
 #' @title Get DisMAP Indicators Table
 #' @description Download species indicators data including Center of Gravity metrics
-#' @param common_name Character string of species common name
-#' @param region Character string of desired region (optional)
-#' @param base_url The base URL for the indicators service
+#' @param ... key = value pairs for filtering; where key could be any of the fields in the indicators table
 #' @return A sf object with indicators data
+#' @importFrom dplyr arrange filter left_join mutate pull select
+#' @importFrom glue glue
+#' @importFrom httr2 request req_perform req_url_query  resp_body_json
+#' @importFrom janitor clean_names
+#' @importFrom jsonlite fromJSON
+#' @importFrom purrr map_chr pluck
+#' @importFrom snakecase to_snake_case to_any_case
+#' @importFrom stringr str_replace
+#' @importFrom tibble tibble
+#'
 #' @export
-get_dismap_indicators <- function(
-    ...,
-    # common_name,
-    # region = NULL,
-    # base_url = "https://services2.arcgis.com/C8EMgrsFcRFL6LrL/ArcGIS/rest/services/Indicators_20220516/FeatureServer/3/query") {
-    # base_url = "https://services2.arcgis.com/C8EMgrsFcRFL6LrL/ArcGIS/rest/services/Indicators_CURRENT/FeatureServer/query") {
-    # base_url = "https://services2.arcgis.com/C8EMgrsFcRFL6LrL/ArcGIS/rest/services/Indicators_CURRENT/FeatureServer/1/query?where=CommonName+%3D+%27American+lobster%27&objectIds=&resultType=none&outFields=*&returnIdsOnly=false&returnUniqueIdsOnly=false&returnCountOnly=false&returnDistinctValues=false&cacheHint=false&collation=&orderByFields=&groupByFieldsForStatistics=&outStatistics=&having=&resultOffset=&resultRecordCount=&sqlFormat=standard&f=pgeojson&token="
-    # utils::URLdecode(base_url)
-    # base_url = "https://services2.arcgis.com/C8EMgrsFcRFL6LrL/ArcGIS/rest/services/Indicators_CURRENT/FeatureServer/1/query?
-    # where=CommonName+=+'American+lobster'&objectIds=&resultType=none&outFields=*&returnIdsOnly=false&returnUniqueIdsOnly=false&returnCountOnly=false&returnDistinctValues=false&cacheHint=false&collation=&orderByFields=&groupByFieldsForStatistics=&outStatistics=&having=&resultOffset=&resultRecordCount=&sqlFormat=standard&f=pgeojson&token="
-    # common_name = "American lobster"; region = NULL;
-    base_url = "https://services2.arcgis.com/C8EMgrsFcRFL6LrL/ArcGIS/rest/services/Indicators_CURRENT/FeatureServer/1/query") {
+get_dm_indicators <- function(
+    ...) {
 
   require(httr2)
-  # require(sf)
   require(jsonlite)
-  # install.packages("arcgisutils", repos = "https://r-arcgis.r-universe.dev")
-  require(arcgisutils) # https://r.esri.com/arcgisutils/
+  if (!(require("arcgisutils"))) {
+    stop("Please install the required arcgis packages with the command: install.packages('arcgis', repos = 'https://r-arcgis.r-universe.dev')")
+  }
   require(stringr)
   require(purrr)
   require(glue)
   require(janitor)
-  # install.packages("RcppSimdJson")
+
+  base_url = "https://services2.arcgis.com/C8EMgrsFcRFL6LrL/ArcGIS/rest/services/Indicators_CURRENT/FeatureServer/1/query"
 
   # get input arguments as list
-  # browser()
   args <- list(...)
 
   # fields
@@ -46,9 +44,8 @@ get_dismap_indicators <- function(
     httr2::resp_body_json() |>
     purrr::pluck("fields") |>
     purrr::map_chr("name")
-  # get name from every element in the f$fields list
-  # f$fields$name
-  d_flds <- tibble(
+
+  d_flds <- tibble::tibble(
     fld_esri  = flds,
     fld_snake = snakecase::to_snake_case(flds),
     fld_camel = snakecase::to_any_case(flds, case = "upper_camel", abbreviations = c("SE","OBJECTID")))
@@ -90,74 +87,117 @@ get_dismap_indicators <- function(
     janitor::clean_names() |>
     dplyr::tibble()
 
-  # Read data
-  # d <- readr::read_csv(req$url)
-  #
-  # sf_indicators <- sf::st_read(req$url, quiet = TRUE)
-  # mapview::mapView(sf_indicators)
-  #
-  # # Convert to data frame
-  # df_indicators <- as.data.frame(sf_indicators)
-
-  # Rename for consistency
-  # indicators_df <- dplyr::rename(indicators_df,
-  #                          lon = CenterOfGravityLongitude,
-  #                          lat = CenterOfGravityLatitude)
-
   return(d)
 }
 
-#' @title Get DisMAP Slice IDs
-#' @description Get slice IDs for a species in a region
-#' @param species_name The scientific name of the species
-#' @param region_url The URL for the image server for this region
-#' @return A data frame with slice information
+#' @title Get DisMAP layer names, given region season
+#' @description Get a list of available layers for a given region and season
+#' @param datset_code The region (and season) code, per available `dataset_code` in `dm_regions`
+#' @return A character vector of layer names, usually species scientific name or "Species Richness"
 #' @export
-get_dismap_slices <- function(species_name, region_url) {
-  require(httr)
-  require(jsonlite)
+#' @importFrom arcgisutils fetch_layer_metadata
+#' @importFrom glue glue
+#' @importFrom purrr pluck
+#' @examples
+#' get_dm_dataset_layers("AI")
+get_dm_dataset_layers <- function(dataset_code){
+  # dataset_code = "AI"
 
-  # URL encode the species name for the API
-  encoded_name <- URLencode(species_name, reserved = TRUE)
+  stopifnot(
+    dataset_code %in% dm_regions$dataset_code,
+    length(dataset_code) == 1)
 
-  # Build the slices URL
-  slices_url <- paste0(
-    region_url,
-    "/slices?multidimensionalDefinition=%5B%7B%22variableName%22%3A+%22",
-    encoded_name,
-    "%22%7D%5D&f=pjson"
-  )
+  glue::glue(
+    "https://maps.fisheries.noaa.gov/image/rest/services/DisMAP/{dataset_code}_IDW_CURRENT/ImageServer/multiDimensionalInfo") |>
+    arcgisutils::fetch_layer_metadata() |>
+    purrr::pluck("multidimensionalInfo", "variables", "name") |>
+    sort()
+}
 
-  # Get the JSON response
-  response <- httr::GET(slices_url)
-  text_json <- httr::content(response, type = 'text', encoding = "UTF-8")
-  jfile <- jsonlite::fromJSON(text_json)
+get_dm_dataset_species_year_slices <- function(dataset_code, species_scientific){
+  # internal function so user doesn't have to worry about sliceId, only needs year
 
-  # Convert to data frame
-  df <- as.data.frame(jfile)
+  # TODO: setup below as function
+  img_url  <- glue::glue("https://maps.fisheries.noaa.gov/image/rest/services/DisMAP/{dataset_code}_IDW_CURRENT/ImageServer")
 
-  return(df)
+  lst_slices <- httr2::request(img_url) |>
+    httr2::req_url_path_append("slices") |>
+    httr2::req_url_query(
+      multidimensionalDefinition = glue("[{{variableName: '{species_scientific}'}}]"),
+      f = "json") |>
+    httr2::req_perform() |>
+    httr2::resp_body_json() |>
+    purrr::pluck("slices")
+
+  tbl_slices <- tibble::tibble(
+    sliceId       = purrr::map_int(lst_slices, "sliceId"),
+    dimensionName = purrr::map_chr(lst_slices, \(x) purrr::pluck(x, "multidimensionalDefinition", 1, "dimensionName")),
+    value         = purrr::map_dbl(lst_slices, \(x) purrr::pluck(x, "multidimensionalDefinition", 1, "values", 1))) |>
+    dplyr::arrange(dimensionName, value)
+
+  stopifnot(all(tbl_slices$dimensionName == "StdTime"))
+
+  tbl_slices <- tbl_slices |>
+    dplyr::mutate(
+      dtime = arcgisutils::from_esri_date(value),
+      year  = lubridate::year(dtime))
+
+  # ensure no duplicate years
+  # (otherwise dtime varying by something other than year)
+  stopifnot(
+    tbl_slices |>
+      dplyr::group_by(year) |>
+      dplyr::summarize(n = dplyr::n()) |>
+      dplyr::filter(n > 1) |>
+      nrow() == 0)
+
+  tbl_slices |>
+    dplyr::select(slice_id = sliceId, year)
+}
+
+#' @title Get DisMAP years for a species
+#' @description Get a list of years for a given species in a DisMAP dataset
+#' @param dataset_code The DisMAP dataset code (e.g., "AI", "EBS", "GOA"), per `dm_regions$dataset_code`
+#' @param species_scientific The scientific name of the species, per `get_dm_dataset_layers(dataset_code)`
+#' @return A vector of years available for the species in the dataset
+#' @export
+#' @importFrom dplyr pull
+#' @examples
+#' get_dm_dataset_species_years("AI", "Paralichthys dentatus")
+get_dm_dataset_species_years <- function(dataset_code, species_scientific){
+  get_dm_dataset_species_year_slices(dataset_code, species_scientific) |>
+    dplyr::pull(year)
 }
 
 #' @title Download DisMAP Raster
 #' @description Download a single interpolated biomass raster by slice ID
-#' @param img_url The URL for the image server for a given region
-#' @param slice_id The slice ID to download
+#' @param dataset_code The DisMAP dataset code (e.g., "AI", "EBS", "GOA"), per `dm_regions$dataset_code`
+#' @param species_scientific The scientific name of the species, per `get_dm_dataset_layers(dataset_code)`
+#' @param year The year of the slice to download, per `get_dm_dataset_species_years(dataset_code, species_scientific)`
 #' @param out_tif optional output GeoTIFF path to write raster
-#' @param transform Function to transform raster values (default is cube root): function(x) x^(1/3)
 #' @param out_sr optional output spatial reference for output raster; defaults to native spatial reference
+#' @param transform Function to transform raster values (default is cube root): function(x) x^(1/3)
 #' @param bbox optional bounding box coordinates (xmin, ymin, xmax, ymax); defaults to maximum extent
 #' @param bbox_sr optional spatial reference for bounding box; defaults to native spatial reference; you can use geographic (4326)
 #' @param overwrite if out_tif exists, overwrite it (default is FALSE)
 #' @param verbose if TRUE, print additional information
 #' @return A terra::rast object
+#' @importFrom httr2 request req_url_query req_perform resp_body_json
+#' @importFrom terra rast writeRaster
+#' @importFrom glue glue
+#' @importFrom janitor clean_names
+#' @importFrom dplyr filter select
+#' @importFrom purrr pluck
+#' @importFrom stringr str_replace_all
+#' @importFrom tibble tibble
 #' @export
-download_dismap_raster <- function(
-    img_url,
-    slice_id,
+get_dm_raster <- function(
+    dataset_code,
+    species_scientific,
+    year,
     out_tif   = NULL,
-    transform = function(x) x^(1/3),
     out_sr    = NULL,
+    transform = function(x) x^(1/3),
     bbox      = NULL,
     bbox_sr   = NULL,
     overwrite = F,
@@ -166,15 +206,40 @@ download_dismap_raster <- function(
   require(httr2)
   require(terra)
 
-  if (file.exists(out_tif) & !overwrite) {
+  if (!is.null(out_tif) && file.exists(out_tif) && !overwrite) {
    if (verbose)
       message("File already exists and overwrite is set to FALSE. Skipping download.")
     r <- terra::rast(out_tif)
     return(r)
   }
 
+  # check for valid dataset_code
+  if (!dataset_code %in% dm_regions$dataset_code)
+    stop("dataset_code must be one of: ", paste(dm_regions$dataset_code, collapse = ", "))
+  if (length(dataset_code) != 1)
+    stop("dataset_code must be a single value")
+
+  # check for valid species_scientific
+  lyrs <- get_dm_dataset_layers(dataset_code)
+  if (!species_scientific %in% lyrs)
+    stop("species_scientific must be one of the dataset's available layers: ", paste(lyrs, collapse = ", "))
+  if (length(species_scientific) != 1)
+    stop("species_scientific must be a single value")
+
+  # check for valid year
+  d_yr_slices <- get_dm_dataset_species_year_slices(dataset_code, species_scientific)
+  yrs <- d_yr_slices$year
+  if (!year %in% yrs)
+    stop("year must be one of the dataset's available years: ", paste(yrs, collapse = ", "))
+  if (!is.numeric(year) || length(year) != 1)
+    stop("year must be a single numeric value")
+  slice_id <- d_yr_slices |>
+    filter(year == !!year) |>
+    pull(slice_id)
+
   if (verbose)
     message("Getting image server information.")
+  img_url  <- glue::glue("https://maps.fisheries.noaa.gov/image/rest/services/DisMAP/{dataset_code}_IDW_CURRENT/ImageServer")
   res <- httr2::request(img_url) |>
     httr2::req_url_query(
       f = "json") |>
@@ -223,7 +288,6 @@ download_dismap_raster <- function(
   names(r) <- "wtcpue"
   # r
   # range(values(r, na.rm = T)) # 0.00000 47.82674
-  # range(values(r2, na.rm = T)) # 0.00000 47.82674
   # plet(r)
 
   if (!is.null(out_tif)) {
@@ -237,255 +301,24 @@ download_dismap_raster <- function(
   r
 }
 
-#' @title Download DisMAP Raster Series
-#' @description Download multiple interpolated biomass rasters for a species over time
-#' @param species_scientific Scientific name of the species
-#' @param species_common Common name of the species (for file naming)
-#' @param region Region name
-#' @param region_url URL for the image server for this region
-#' @param bbox Bounding box coordinates (xmin, ymin, xmax, ymax)
-#' @param output_dir Directory to save rasters
-#' @param years Optional vector of years to download (if NULL, downloads all available years)
-#' @return Data frame with paths to downloaded rasters
-#' @export
-download_dismap_series <- function(species_scientific,
-                                  species_common,
-                                  region,
-                                  region_url,
-                                  bbox = NULL,
-                                  output_dir = ".",
-                                  years = NULL) {
-
-  require(dplyr)
-
-  # Create output directory if it doesn't exist
-  dir_path <- file.path(output_dir, paste0(species_common, "_", gsub(" ", "_", region)))
-  if (!dir.exists(dir_path)) {
-    dir.create(dir_path, recursive = TRUE)
-  }
-
-  # Get slice IDs for the species
-  slice_data <- get_dismap_slices(species_scientific, region_url)
-
-  # Filter by years if provided
-  if (!is.null(years)) {
-    # This assumes slice metadata contains a key for the year
-    # May need adjustment based on actual API response
-    slice_ids <- slice_data %>%
-      dplyr::filter(keyValue %in% as.character(years)) %>%
-      dplyr::pull(id)
-  } else {
-    slice_ids <- slice_data$slices.id
-  }
-
-  # Download each raster
-  results <- data.frame(
-    slice_id = integer(),
-    year = integer(),
-    file_path = character(),
-    stringsAsFactors = FALSE
-  )
-
-  for (i in seq_along(slice_ids)) {
-    slice_id <- slice_ids[i]
-
-    # Get year from slice data
-    year <- slice_data$slices.keyValue[slice_data$slices.id == slice_id]
-
-    # Output file path
-    file_name <- paste0(species_common, "_", year, "_", slice_id)
-    file_path <- file.path(dir_path, file_name)
-
-    # Download raster
-    message(sprintf("Downloading raster for %s, year %s (slice %s)", species_common, year, slice_id))
-
-    download_dismap_raster(
-      slice_id = slice_id,
-      region_url = region_url,
-      bbox = bbox,
-      out_file = file_path
-    )
-
-    # Add to results
-    results <- rbind(results, data.frame(
-      slice_id = slice_id,
-      year = as.integer(year),
-      file_path = file_path,
-      stringsAsFactors = FALSE
-    ))
-  }
-
-  return(results)
-}
-
-
-#' @title Plot DisMAP Distribution Map
-#' @description Create a static or animated plot of species distribution
-#' @param data Data frame with lon, lat, year, and transformed biomass columns
-#' @param value_col Name of column with values to plot
-#' @param animated Whether to create an animated plot
-#' @param xlim,ylim Limits for x and y axes
-#' @param color_palette Color palette function
-#' @return A ggplot object or animated gif
-#' @export
-plot_dismap_distribution <- function(data,
-                                    value_col = "transformed_wtcpue",
-                                    animated = FALSE,
-                                    xlim = NULL,
-                                    ylim = NULL,
-                                    color_palette = NULL) {
-
-  require(ggplot2)
-  require(maps)
-
-  # Set defaults for limits if not provided
-  if (is.null(xlim)) {
-    xlim <- c(min(data$lon, na.rm = TRUE), max(data$lon, na.rm = TRUE))
-  }
-
-  if (is.null(ylim)) {
-    ylim <- c(min(data$lat, na.rm = TRUE), max(data$lat, na.rm = TRUE))
-  }
-
-  # Set default color palette if not provided
-  if (is.null(color_palette)) {
-    if (requireNamespace("cmocean", quietly = TRUE)) {
-      color_palette <- cmocean::cmocean("matter")(256)
-    } else {
-      color_palette <- viridis::viridis(256)
-    }
-  }
-
-  # Create base plot
-  p <- ggplot(data = data, aes(x = lon, y = lat)) +
-    geom_tile(aes_string(fill = value_col)) +
-    theme_classic() +
-    labs(y = "", x = "") +
-    theme(legend.position = "right", legend.title = element_blank()) +
-    theme(panel.border = element_rect(colour = "black", fill = NA, size = 1)) +
-    scale_fill_gradientn(colours = color_palette,
-                          limits = c(0, max(data[[value_col]], na.rm = TRUE))) +
-    annotation_map(map_data("world"), colour = "black", fill = "grey50") +
-    coord_quickmap(xlim = xlim, ylim = ylim) +
-    scale_x_continuous(expand = c(0, 0)) +
-    scale_y_continuous(expand = c(0, 0))
-
-  # Add animation if requested
-  if (animated) {
-    if (!requireNamespace("gganimate", quietly = TRUE)) {
-      stop("Package 'gganimate' needed for animation. Please install it.")
-    }
-
-    p <- p +
-      gganimate::transition_time(year) +
-      gganimate::ease_aes("linear") +
-      labs(title = "{frame_time}")
-
-    # Render animation
-    if (requireNamespace("gifski", quietly = TRUE)) {
-      anim <- gganimate::animate(p,
-                               nframes = length(unique(data$year)),
-                               fps = 2,
-                               renderer = gganimate::gifski_renderer())
-      return(anim)
-    } else {
-      warning("Package 'gifski' is recommended for creating GIFs. Using default renderer.")
-      anim <- gganimate::animate(p,
-                               nframes = length(unique(data$year)),
-                               fps = 2)
-      return(anim)
-    }
-  } else {
-    return(p)
-  }
-}
-
 #' @title Get DisMAP Survey Locations
 #' @description Download survey location points for a region
-#' @param region Region name
-#' @param base_url Base URL for survey locations service
+#' @param dataset_code The DisMAP dataset code (e.g., "AI", "EBS", "GOA"), per `dm_regions$dataset_code`
+#' @param where Optional SQL WHERE clause to filter results (e.g., "year = 2015")
 #' @return A sf object with survey locations
 #' @export
-get_survey_locations <- function(
-    region,
-    base_url = "https://services2.arcgis.com/C8EMgrsFcRFL6LrL/ArcGIS/rest/services") {
+get_dm_survey_locations <- function(
+    dataset_code,
+    where = NULL){
 
-  require(httr2)
-  require(sf)
+  url <- glue("https://services2.arcgis.com/C8EMgrsFcRFL6LrL/arcgis/rest/services/{dataset_code}_Sample_Locations_CURRENT/FeatureServer")
+  fs <- arcgislayers::arc_open(url)
+  lyr <- arcgislayers::get_layer(fs, 1)
 
-
-  # Build URL
-  d <- httr2::request(base_url) |>
-    httr2::req_url_query(
-      where     = where_clause,
-      outFields = "*",
-      f         = "json") |>
-    httr2::req_perform() |>
-    httr2::resp_body_string() |>
-    arcgisutils::parse_esri_json() |>
-    janitor::clean_names()
-
-  # Extract region code from name for URL building
-  region_code <- gsub(" ", "_", region)
-
-  # Build URL
-  url <- httr::parse_url(base_url)
-  url$path <- paste(url$path, paste0(region_code, "_Survey_Locations_20220516/FeatureServer/1/query"), sep = "/")
-  url$query <- list(
-    where = sprintf("OARegion='%s'", region),
-    outFields = "*",
-    f = "geojson"
-  )
-
-  request <- httr::build_url(url)
-
-  # Read data
-  survey_locations <- sf::st_read(request, quiet = TRUE)
-
-  return(survey_locations)
+  arcgislayers::arc_select(
+    lyr,
+    where = where) |>
+    tibble::tibble() |>
+    sf::st_as_sf()
 }
 
-#' @title List Available DisMAP Regions
-#' @description Return a list of available regions in DisMAP
-#' @return Character vector of region names
-#' @export
-list_dismap_regions <- function() {
-  regions <- c(
-    "Aleutian Islands",
-    "Eastern Bering Sea",
-    "Gulf of Alaska",
-    "West Coast Triennial",
-    "West Coast Annual",
-    "Gulf of Mexico",
-    "Northeast US Spring",
-    "Northeast US Fall",
-    "Southeast US Fall",
-    "Southeast US Spring",
-    "Southeast US Summer",
-    "Northern Bering Sea",
-    "Eastern and Northern Bering Sea"
-  )
-  return(regions)
-}
-
-#' @title Get DisMAP Base URLs
-#' @description Get the base URLs for various DisMAP services
-#' @return A list with URLs for indicators, surveys, and rasters
-#' @export
-get_dismap_urls <- function() {
-  urls <- list(
-    indicators = "https://services2.arcgis.com/C8EMgrsFcRFL6LrL/arcgis/rest/services/Indicators_20220516/FeatureServer",
-    surveys = "https://services2.arcgis.com/C8EMgrsFcRFL6LrL/ArcGIS/rest/services",
-    rasters = list(
-      "Northeast US Spring" = "https://maps.fisheries.noaa.gov/image/rest/services/DisMAP/Northeast_US_Spring_20220516/ImageServer",
-      "Northeast US Fall" = "https://maps.fisheries.noaa.gov/image/rest/services/DisMAP/Northeast_US_Fall_20220516/ImageServer",
-      # Add other regions as needed
-      "Eastern Bering Sea" = "https://maps.fisheries.noaa.gov/image/rest/services/DisMAP/Eastern_Bering_Sea_20220516/ImageServer"
-    )
-  )
-
-  # Add note about checking inPort for latest URLs
-  message("NOTE: Check https://www.fisheries.noaa.gov/inport/item/66799 for the most up-to-date URLs")
-
-  return(urls)
-}